+++ lib/flutter_elink_ble.dart
@@ -1,3 +1,3 @@
 export 'src/elink_ble.dart';
 export 'src/elink_data_processor.dart';
-export 'src/elink_models.dart';
+export 'src/elink_ble_models.dart';